Adding animation test files from Sunday
This is the status they were in when I opened them late Monday night. I
have not messed with them since the last you tried on Sunday, Keith, so
have at it. I will try running them now to see what they do. I think we
need to learn how to operate the Chip Select lines on the WiFi shield,
so we can use the CC3000 and the SD card in the same code. Still lots
to do, but we are getting there!